+++ backend/src/features/memo/model/request/updateMemo.go
@@ -8,7 +8,7 @@ type ReqUpdateMemo struct {
 	ImageURL        string                `json:"image_url"`
 	ImageFile       multipart.File        `json:"-"` // S3 업로드용 파일
 	ImageHeader     *multipart.FileHeader `json:"-"` // 파일 메타데이터
-	Rating          uint8                 `json:"rating"`
+	Rating          float32               `json:"rating"` // float로 받아서 uint8로 변환
 	IsPinned        bool                  `json:"is_pinned"`
 	Latitude        *float64              `json:"latitude"`
 	Longitude       *float64              `json:"longitude"`
